% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria_mseL.R
\name{criteria.mseL}
\alias{criteria.mseL}
\title{Calculates the values of the MSE LPs-criterion and its components}
\usage{
criteria.mseL(X1, X2, search.object, eps = 10^-23)
}
\arguments{
\item{X1}{The primary model matrix, with the first column containing the labels of treatments, and the second -- the intercept term.}

\item{X2}{The matrix of potential terms, with the first column containing the labels of treatments.}

\item{eps}{Computational tolerance, the default value is 10^-23}
}
\value{
A list of values: indicator of whether the evaluation was successful ("eval"), LPs-criterion value -- intercept excluded ("LP"),
Lack-of-fit(LP) criterion value ("LoF"), the MSE(L) component value ("mse"),
the number of pure error degrees of freedom ("df") and the value of the compound criterion ("compound").
}
\description{
This function evaluates the MSE LPs-criterion for given primary and potential model matrices. Candidate full model matrices do not have to be orthonormalised.
Components: LPs-, LoF(LP)- and MSE(L)-optimality.
The weights kappa.LP, kappa.LoF and kappa.mse, and other parameters (tau2, alpha-s) are taken from the global environment.
}
\examples{
# Experiment: one 5-level factor, primary model -- full quadratic, X^3 and X^4 potential terms.
K<-1; P<-3; Q<-2;
Levels <- list(1:5)   
# Generating candidate sets: primary and full ones
cand.primary <- candidate_set(Levels);
cand.full <- cbind(cand.primary, cand.primary[,3]^3, cand.primary[,4]^2) #X^3 and X^4 potential terms
# Choosing a design
index <- c(rep(1,2),3,rep(4,2),rep(5,3)); Nruns<- length(index)
Z0<-diag(1,Nruns)-matrix(1/Nruns,nrow=Nruns,ncol=Nruns)
X.primary <- cand.full[index, 1:(P+1)]
X.potential <- cand.full[index, (c(1,(P+2):(P+Q+1)))]
# Evaluating a compound MSE(LP)-criterion
kappa.LP = kappa.LoF = kappa.mse = 1./3; tau2 <-1; tau <- sqrt(tau2)
alpha.LP = alpha.LoFL = 0.05;  
criteria.mseL(X1 = X.primary, X2 = X.potential)
# Output: eval = 1, LP = 2.3863, LoF = 7.1846, mse = 1.5994, df = 4, compound = 2.8047

}
