% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{criteria}
\alias{criteria}
\title{Evaluates the design based on a specific optimality criterion}
\usage{
criteria(X1, X2, criterion.choice)
}
\arguments{
\item{criterion.choice}{Criterion choice can be "GL", "GLP", "GD", "GDP", "MSE.L", "MSE.D", "MSE.P".}
}
\value{
A list...
}
\description{
This function evaluates the design based on a specific optimality criterion.
}
