% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOODE-package.R
\docType{package}
\name{MOODE-package}
\alias{MOODE}
\alias{MOODE-package}
\title{MOODE: Multi-objective optimal design of experiments}
\description{
Multi-objective Optimal Design of Experiments (MOODE) for targeting the experimental objectives directly, ensuring as such that the full set of research questions is answered as economically as possible.
}
\keyword{internal}
