% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objfun.R
\name{objfun}
\alias{objfun}
\title{Evaluates the design based on a specific optimality criterion}
\usage{
objfun(X1, X2, search.object)
}
\arguments{
\item{criterion.choice}{Criterion choice can be "GL", "GLP", "GD", "GDP", "MSE.L", "MSE.D", "MSE.P".}
}
\value{
A list...
}
\description{
This function evaluates the design based on a specific optimality criterion.
}
