% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointexchange.R
\name{point.swap}
\alias{point.swap}
\title{Swapping points between the current design and candidate set}
\usage{
point.swap(X1, X2, cand.full, search.object)
}
\arguments{
\item{X1}{Current fitted (primary) model matrix}

\item{X2}{Current potential terms matrix}

\item{cand.full}{Full candidate matrix}

\item{search.object}{Object for the search}
}
\value{
A list of model matrices, criteria values and
whether the search needs to continue
}
\description{
Performing point-exchange algorithm, extensive swap of points procedure between the current design
and candidate set.
}
\details{
\code{point.swap} is called within the \code{Search} function
}
\keyword{internal}
