% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.settings.R
\name{print.settings}
\alias{print.settings}
\title{S3 print method}
\usage{
\method{print}{settings}(x, ...)
}
\arguments{
\item{x}{mood object}

\item{...}{further arguments passed to or from other methods}
}
\description{
S3 print method
}
