% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Search.R
\name{Search}
\alias{Search}
\title{Searching for a multi-objective optimal completely randomised design.}
\usage{
Search(mood.object, pointex = TRUE, update.info = TRUE)
}
\arguments{
\item{mood.object}{The object generated by \code{mood} function, containing the parameters of the experiment, the compound criterion and search parameters}

\item{pointex}{Parameter specifying the search algorithm. If \code{TRUE}, the point-exchange algorithm is used, and if \code{FALSE} -- the coordinate-exchange}

\item{update.info}{If \code{TRUE}, progress messages through the search iterations are shown.}
}
\value{
List of the outputs generated by the search:
\itemize{
\item \code{X.design} Design matrix.
\item \code{df} The number of pure error degrees of freedom.
\item \code{X1} Primary model matrix for the found (nearly-) optimum design.
\item \code{X2} Model matrix of potential terms for the found (nearly-) optimum design.
\item \code{compound.value} The compound criterion value of the (nearly-) optimum design.
\item \code{criteria.values} Component criteria values of the (nearly-) optimum design.
\item \code{path} The "path" of compound criterion values of the optimum designs obtained after for each random start.
\item \code{time} Computation time.
}
}
\description{
Performing search for a (nearly) optimum factorial design, oprimising with respect to a specified compound criterion.
}
\details{
\code{Search} takes the mood object as an input with all the parameters of the experiment. Runs a point-exchange or a coordinate-exchange algorithm, returns design and model matrices, computation time and criteria values.
}
\examples{

example1 <- mood(K = 5, Klev = 3, Nruns = 40, criterion.choice = "GDP", 
                kappa.Ds = 1./3, kappa.DP = 1./3, kappa.LoF = 1./3, 
                Nstarts = 50, tau2 = 0.1, primary.model = "second_order",
                potential.model = NA, 
                potential.terms = c("x12x2", "x22x3", "x32x4", "x42x5"))
example1

Search_point <- Search(example1, pointex = TRUE)
Search_point
Search_coord <- Search(example1, pointex = FALSE)
Search_coord
}
\seealso{
\link{mood}
}
