% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{settings}
\alias{settings}
\title{Setting up the parameters of the experiment, criteria and search procedure}
\usage{
settings(
  K,
  Klev,
  Nruns = 40,
  criterion.choice = "MSE.P",
  kappa.DP = 1,
  kappa.LoF = 0,
  kappa.mse = 0,
  kappa.Ds = 0,
  kappa.Ls = 0,
  Nstarts = 10,
  Biter = 50,
  Cubic = "Y",
  tau2 = 1,
  MC = "Y",
  prob.LP = 0.95,
  prob.LoFL = 0.95,
  alpha.LoF = 0.05,
  alpha.DP = 0.05,
  orth = "N"
)
}
\arguments{
\item{K}{}
}
\value{
The object
}
\description{
This function
}
\examples{

 


}
