% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_set.R
\name{candidate_set}
\alias{candidate_set}
\title{Forms the candidate set}
\usage{
candidate_set(Levels)
}
\arguments{
\item{Levels}{Levels of each factor.}
}
\value{
The candidate set.
}
\description{
This function forms the candidate set as all possible combinations of factors levels.
}
\examples{
candidate_set(rep(list(1:3),5)) # Candidate set of 5 factors each of 3 levels
}
