% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_set_full.R
\name{candidate_set_full}
\alias{candidate_set_full}
\title{Forms the full extended candidate set}
\usage{
candidate_set_full(cand)
}
\arguments{
\item{cand}{Candidate set.}
}
\value{
The full extended candidate set.
}
\description{
This function forms the full extended candidate set, primary terms, with labels, not orthonormalised.
}
\examples{
candidate_set_full(candidate_set(rep(list(1:3),5))) # Full extended candidate set of 5 factors each of 3 levels
}
