% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_set_full.R
\name{candidate_set_full}
\alias{candidate_set_full}
\title{Forms the full candidate set of treatments polynomial terms}
\usage{
candidate_set_full(cand, K)
}
\arguments{
\item{cand}{Candidate set of treatments, the first column contains treatment labels. 
Usually obtained as output from the "candidate_set" function.}
}
\value{
The full extended candidate set: the column of treatment labels, then named columns with polynomial terms up to the order of 3. 
For example, "x12" stands for $x_1^2$, and "x1x2 stands for $x_1x_2$, and "x23x4" -- for $x_2^3x_4$.
}
\description{
This function forms the full extended candidate set with all polynomial terms, with labels, not orthonormalised.
}
\examples{

# Full extended candidate set for two 3-level factors

K<-2; Levels <- rep(list(1:3),K);
candidate_set_full(candidate_set(Levels, K), K)
}
