% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_set_full.R
\name{candidate_set_full}
\alias{candidate_set_full}
\title{Forms the full candidate set}
\usage{
candidate_set_full(cand)
}
\arguments{
\item{cand}{Candidate set of primary terms, the first column contains treatment labels. 
Usually obtained as output from the "candidate_set" function.}
}
\value{
The full extended candidate set: the column of treatment labels, then P primary terms, then Q potential terms.
}
\description{
This function forms the full extended candidate set by joining primary and potential terms, with labels, not orthonormalised.
}
\examples{

# Full extended candidate set for two 3-level factors, full quadratic polynomial model

K<-2; Levels <- rep(list(1:3),K);
Parameters <- c(1, rep(1,K), rep(1,K), K*(K-1)/2)
candidate_set_full(candidate_set(Levels))
}
