% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_set_orth.R
\name{candidate_set_orth}
\alias{candidate_set_orth}
\title{Forms the orthonormalised full candidate set}
\usage{
candidate_set_orth(cand)
}
\arguments{
\item{cand}{Candidate set containing primary terms, with labels in the first column.}
}
\value{
The orthonormalised full candidate set containing primary and potential terms, with labels.
}
\description{
This function forms the full extended orthonormalised candidate set of primary and potential terms, with labels.
}
\examples{

# Full extended orthonormalised candidate set for two 3-level factors, 
# full quadratic polynomial model as primary model and all three-order terms as potential.

K<-2; Levels <- rep(list(1:3),K);
Parameters <- c(1, rep(1,K), rep(1,K), K*(K-1)/2) 
candidate_set_orth(candidate_set(Levels)) 
}
