% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria_GD.R
\name{criteria.GD}
\alias{criteria.GD}
\title{Calculates the values of the Generalised Ds-criterion and its components}
\usage{
criteria.GD(X1, X2, search.object, eps = 10^-23)
}
\arguments{
\item{X1}{The primary model matrix, with the first column containing the labels of treatments, and the second -- the intercept term.}

\item{X2}{The matrix of potential terms, with the first column containing the labels of treatments.}

\item{eps}{Computational tolerance, the default value is 10^-23}
}
\value{
A list of values: indicator of whether the evaluation was successful ("eval"), Ds-criterion value -- intercept excluded ("Ds"),
Lack-of-fit criterion value ("LoF"), the bias component value ("bias"), the number of pure error degrees of freedom ("df")
and the value of the compound criterion ("compound").
}
\description{
This function evaluates the Generalised Ds-criterion (Goos et al., 2005) for given primary and potential model matrices.
The weights kappa.Ds, kappa.LoF and kappa.bias are taken from the global environment.
}
\examples{
#Experiment: one 5-level factor, primary model -- full quadratic, one potential (cubic) term
K <-1; P<-3; Q<-1; Levels <- list(1:5)
# Generating candidate sets: primary and full orthonormalised ones
cand.primary <- candidate_set(Levels);
cand.not.orth <-cbind(cand.primary[,-1], cand.primary[,3]*cand.primary[,4])
cand.full.orth <- cbind(cand.primary[,1], far::orthonormalization(cand.not.orth,basis=FALSE))
# Choosing a design
index <- c(rep(1,2),3,4, rep(5,3)); Nruns<- length(index)
X.primary <- cand.full.orth[index, 1:(P+1)]
X.potential <- cand.full.orth[index, (c(1,(P+2):(P+Q+1)))]
# Evaluating a compound GD-criterion
kappa.Ds = kappa.LoF = kappa.bias = 1./3; tau2 <-1;
criteria.GD(X1 = X.primary, X2 = X.potential)
# Output: eval = 1, Ds = 1.4045, LoF = .7213, bias = 1.4331, df = 3, compound = 1.1427
}
