% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria_GL.R
\name{criteria.GL}
\alias{criteria.GL}
\title{Calculates the values of the Generalised Ls-criterion and its components}
\usage{
criteria.GL(X1, X2, eps = 10^-23)
}
\arguments{
\item{X1}{The primary model matrix, with the first column containing the labels of treatments, and the second -- the intercept term.}

\item{X2}{The matrix of potential terms, with the first column containing the labels of treatments.}

\item{eps}{Computational tolerance, the default value is 10^-23}
}
\value{
A list of values: indicator of whether the evaluation was successful ("eval"), Ls-criterion value -- intercept excluded ("Ls"),
Lack-of-fit criterion value ("LoF"), the bias component value ("bias"), the number of pure error degrees of freedom ("df") 
and the value of the compound criterion ("compound").
}
\description{
This function evaluates the Generalised Ls-criterion (Goos et al., 2005) for given primary and potential model matrices.
The weights kappa.Ls, kappa.LoF and kappa.bias are taken from the global environment.
}
\examples{
kappa.Ls = kappa.LoF = kappa.bias = 1./3;
cand.primary <- candidate_set(rep(list(1:3), 2));
index <- c(1,2,4,6,7); P <- 6; Q <- 2;
X.primary <- cand.primary[index,1:(P+1)]
X.potential <- cbind(X.primary[,1], X.primary[,2]^3, X.primary[,3]^3) # cubic terms
criteria.GL(X1 = X.primary, X2 = X.potential)
}
