% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria_GLP.R
\name{criteria.GLP}
\alias{criteria.GLP}
\title{Calculates the values of the Generalised LPs-criterion and its components}
\usage{
criteria.GLP(X1, X2, eps = 10^-23)
}
\arguments{
\item{X1}{The primary model matrix, with the first column containing the labels of treatments, and the second -- the intercept term.}

\item{X2}{The matrix of potential terms, with the first column containing the labels of treatments.}

\item{eps}{Computational tolerance, the default value is 10^-23}
}
\value{
A list of values: indicator of whether the evaluation was successful ("eval"), Ls-criterion value -- intercept excluded ("Ls"),
LPs-criterion value -- intercept excluded ("LPs"), Lack-of-fit(LP) criterion value ("LoF"), the bias component value ("bias"), 
the number of pure error degrees of freedom ("df") and the value of the compound criterion ("compound").
}
\description{
This function evaluates the Generalised LPs-criterion for given primary and potential model matrices. 
Components: Ls-, LPs-, LoF(LP)- and Bias(L)-optimality.
The weights kappa.Ls, kappa.LP, kappa.LoF and kappa.bias are taken from the global environment.
}
\examples{
#Experiment: one 5-level factor, primary model -- full quadratic, X^3 and X^4 potential terms.
K <-1; P<-3; Q<-2; Levels <- list(1:5)
# Generating candidate sets: primary and full orthonormalised ones
cand.primary <- candidate_set(Levels);
cand.not.orth <-cbind(cand.primary[,-1], cand.primary[,3]^3, cand.primary[,4]^2)
cand.full.orth <- cbind(cand.primary[,1], far::orthonormalization(cand.not.orth,basis=FALSE))
# Choosing a design
index <- c(rep(1,2),3,rep(4,2),rep(5,3)); Nruns<- length(index)
X.primary <- cand.full.orth[index, 1:(P+1)]
X.potential <- cand.full.orth[index, (c(1,(P+2):(P+Q+1)))]
# Evaluating a compound GLP-criterion
kappa.Ls = kappa.LP = kappa.LoF = kappa.bias = 0.25; tau2 <-1;
alpha.LP = alpha.LoFL = 0.05;
criteria.GLP(X1 = X.primary, X2 = X.potential)

Output: eval = 1, Ls = .5315, LP = 4.0969, LoF = 5.3727, bias = 1.4013, df = 4, compound = 2.0122

}
