% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOODE-package.R
\docType{package}
\name{MOODE-package}
\alias{MOODE}
\alias{MOODE-package}
\title{MOODE: Multi-Objective Optimal Design of Experiments}
\description{
Provides functionality to generate compound optimal designs for targeting the multiple experimental objectives directly, ensuring that the full set of research questions is answered as economically as possible. Designs can be found using point or coordinate exchange algorithms combining estimation, inference and lack-of-fit criteria that account for model inadequacy. Details and examples are given by Koutra et al. (2024) <doi.org/10.48550/arXiv.2412.17158>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vkstats/MOODE}
  \item Report bugs at \url{https://github.com/vkstats/MOODE/issues}
}

}
\author{
\strong{Maintainer}: Vasiliki Koutra \email{vasiliki.koutra@kcl.ac.uk} (\href{https://orcid.org/0000-0002-1117-4554}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Olga Egorova [copyright holder]
  \item Steven Gilmour [copyright holder]
  \item Luzia Trinca [copyright holder]
}

}
\keyword{internal}
